change package of the libraries
