+++ dynamicadapter/proguard-rules.pro
