+++ dynamicadapter/src/main/java/me/thuongle/dynamicadapter/handler/ChildItemsClickBinder.java
@@ -1,0 +1,10 @@
+package me.thuongle.dynamicadapter.handler;
+
+import java.util.List;
+
+public interface ChildItemsClickBinder<T> {
+
+    int getBindingVariable();
+
+    List<ClickHandler<T>> clickHandlers();
+}
