+++ dynamicadapter/src/main/java/me/thuongle/dynamicadapter/handler/ClickHandler.java
@@ -1,0 +1,5 @@
+package me.thuongle.dynamicadapter.handler;
+
+public interface ClickHandler<T> {
+    void onClick(T model);
+}
