+++ dynamicadapter/src/main/java/me/thuongle/dynamicadapter/handler/ItemBinder.java
@@ -1,0 +1,8 @@
+package me.thuongle.dynamicadapter.handler;
+
+public interface ItemBinder {
+
+    int getBindingVariable();
+
+    int getLayoutRes();
+}
