+++ dynamicadapter/src/main/java/me/thuongle/dynamicadapter/handler/ItemTouchHandler.java
@@ -1,0 +1,8 @@
+package me.thuongle.dynamicadapter.handler;
+
+public interface ItemTouchHandler<T> {
+
+    void onItemMove(int position, T model);
+
+    void onItemDismiss(int position, T model);
+}
