+++ dynamicadapter/src/main/java/me/thuongle/dynamicadapter/handler/LongClickHandler.java
@@ -1,0 +1,5 @@
+package me.thuongle.dynamicadapter.handler;
+
+public interface LongClickHandler<T> {
+    void onLongClick(T model);
+}
