+++ dynamicadapter/src/main/java/me/thuongle/dynamicadapter/handler/OnItemTouchListener.java
@@ -1,0 +1,10 @@
+package me.thuongle.dynamicadapter.handler;
+
+import android.view.View;
+
+public interface OnItemTouchListener {
+
+    boolean onItemMove(View v, int fromPosition, int toPosition);
+
+    boolean onItemDismiss(View v, int position);
+}
