+++ dynamicadapter/src/main/res/values/attrs.xml
