+++ exandabledynamicadapter/build.gradle
