+++ exandabledynamicadapter/proguard-rules.pro
