+++ exandabledynamicadapter/src/main/res/values/strings.xml
