+++ exandabledynamicadapter/src/test/java/me/thuongle/exandabledynamicadapter/ExampleUnitTest.java
@@ -1,0 +1,15 @@
+package me.thuongle.exandabledynamicadapter;
+
+import org.junit.Test;
+
+import static org.junit.Assert.*;
+
+/**
+ * To work on unit tests, switch the Test Artifact in the Build Variants view.
+ */
+public class ExampleUnitTest {
+    @Test
+    public void addition_isCorrect() throws Exception {
+        assertEquals(4, 2 + 2);
+    }
+}
