--- exandablerecycleradapter/src/androidTest/java/me/thuongle/exandablerecycleradapter/ApplicationTest.java
@@ -1,13 +1,0 @@
-package me.thuongle.exandablerecycleradapter;
-
-import android.app.Application;
-import android.test.ApplicationTestCase;
-
-/**
- * <a href="http://d.android.com/tools/testing/testing_android.html">Testing Fundamentals</a>
- */
-public class ApplicationTest extends ApplicationTestCase<Application> {
-    public ApplicationTest() {
-        super(Application.class);
-    }
-}
