--- recycleradapter/src/main/AndroidManifest.xml
@@ -1,5 +1,0 @@
-<manifest package="me.thuongle.recycleradapter">
-
-    <application />
-
-</manifest>
