--- recycleradapter/src/main/java/me/thuongle/recycleradapter/handler/ItemBinder.java
@@ -1,8 +1,0 @@
-package me.thuongle.recycleradapter.handler;
-
-public interface ItemBinder {
-
-    int getBindingVariable();
-
-    int getLayoutRes();
-}
