--- recycleradapter/src/main/java/me/thuongle/recycleradapter/handler/LongClickHandler.java
@@ -1,5 +1,0 @@
-package me.thuongle.recycleradapter.handler;
-
-public interface LongClickHandler<T> {
-    void onLongClick(T model);
-}
