--- recycleradapter/src/main/java/me/thuongle/recycleradapter/handler/OnItemTouchListener.java
@@ -1,10 +1,0 @@
-package me.thuongle.recycleradapter.handler;
-
-import android.view.View;
-
-public interface OnItemTouchListener {
-
-    boolean onItemMove(View v, int fromPosition, int toPosition);
-
-    boolean onItemDismiss(View v, int position);
-}
