+++ sample/build.gradle
@@ -35,8 +35,8 @@ dependencies {
     compile 'com.android.support:recyclerview-v7:23.4.0'
     compile 'com.android.support:cardview-v7:23.4.0'
 
-    compile project(':recycleradapter')
-    compile project(':exandablerecycleradapter')
+    compile project(':dynamicadapter')
+    compile project(':exandabledynamicadapter')
 
     compile 'io.reactivex:rxjava:1.1.3'
     compile 'io.reactivex:rxandroid:1.1.0'
